add tests for actions and reducers.
