+++ src/_tests_/actions/user.test.js
@@ -1,0 +1,9 @@
+import addUserId from '../../actions/user';
+
+test('should set up to add user id', () => {
+  const action = addUserId(1);
+  expect(action).toEqual({
+    type: 'ADD_USER_ID',
+    id: 1,
+  });
+});
