+++ src/_tests_/components/App.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import toJSON from 'enzyme-to-json';
-import App from '../components/App';
+import App from '../../components/App';
 
 Enzyme.configure({ adapter: new Adapter() });
