+++ src/_tests_/components/AppointmentList.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import Home from '../components/Home';
+import AppointmentList from '../../components/AppointmentList';
 
 Enzyme.configure({ adapter: new Adapter() });
 
 test('should render AppointmentList component', () => {
-  const wrapper = shallow(<Home />);
+  const wrapper = shallow(<AppointmentList />);
   expect(wrapper).toMatchSnapshot();
 });
