+++ src/_tests_/components/Home.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import Menu from '../components/Menu';
+import Home from '../../components/Home';
 
 Enzyme.configure({ adapter: new Adapter() });
 
-test('should render Menu component', () => {
-  const wrapper = shallow(<Menu />);
+test('should render AppointmentList component', () => {
+  const wrapper = shallow(<Home />);
   expect(wrapper).toMatchSnapshot();
 });
