+++ src/_tests_/components/Menu.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import AppointmentList from '../components/AppointmentList';
+import Menu from '../../components/Menu';
 
 Enzyme.configure({ adapter: new Adapter() });
 
-test('should render AppointmentList component', () => {
-  const wrapper = shallow(<AppointmentList />);
+test('should render Menu component', () => {
+  const wrapper = shallow(<Menu />);
   expect(wrapper).toMatchSnapshot();
 });
