+++ src/_tests_/components/Physician.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import { Physician } from '../components/Physician';
+import { Physician } from '../../components/Physician';
 
 Enzyme.configure({ adapter: new Adapter() });
